+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
     "objects",
     "bin/bench-tx",
     "bin/tx-prover",
-    "bin/prover-proxy"
 ]
 
 [workspace.package]
