--- bin/prover-proxy/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "miden-prover-proxy"
-version = "0.6.0"
-description = "Miden rollup transaction prover proxy"
-readme = "README.md"
-keywords = ["miden", "transaction", "prover", "proxy"]
-license.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-rust-version.workspace = true
-edition.workspace = true
-
-[[bin]]
-name = "miden-prover-proxy"
-path = "src/main.rs"
-
-[dependencies]
-async-trait="0.1"
-pingora = { version = "0.3", features = [ "lb" ] }
-pingora-core = "0.3"
-pingora-proxy = "0.3"
-pingora-limits = "0.3"
-once_cell = "1.19.0"
-tracing-subscriber = "0.3"
-rand = "0.8"
-tokio = { version = "1", features = ["full"] }
-tracing = "0.1"
