+++ bin/tx-prover/src/worker_main.rs
